requirements.txt
# FastAPI and related
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4

# Utilities
pydantic==1.10.7
pydantic[email]==1.10.7

# CSV handling
pandas==2.0.1

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0